+++ R/methods__userinfo.R
@@ -41,7 +41,9 @@ get_userinfo <- function(
 
   if (S7::S7_inherits(token, class = OAuthToken)) {
     access_token <- token@access_token
-    token_type <- token@token_type %||% token_type
+    if (is_valid_string(token@token_type)) {
+      token_type <- token@token_type
+    }
   } else {
     access_token <- token
   }